backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
chromadb>=0.5.23
pydantic==2.7.1
python-multipart==0.0.9
openai>=1.0.0
astral>=3.2
python-dotenv>=1.0.0